happy_goat_time/requirements.txt
#Django
Django==1.7.7
# Templates
jingo==0.7

# Froms
git+http://github.com/maraujop/django-crispy-forms.git@dev

# Celery: Message queue
celery==3.1.17
django-celery==3.1.16
# Feature testing
django-waffle==0.10.1

#debugging
# if you have issues on mavericks installing pycrypto try:
# ARCHFLAGS=-Wno-error=unused-command-line-argument-hard-error-in-future pip install pycrypto
ipython==3.0.0
#ipython==2.4.1
ipdb
Werkzeug==0.10.4
#Werkzeug==0.9.6
django-debug-toolbar==1.3.0

#misc
#wsgiref==0.1.2
django-extensions==1.5.2
#django-extensions==1.5.0
django-braces==1.4.0
django-betterforms==1.1.1
# django-extra-views==0.6.5
# django-organizations==0.4.1
# install the dev version of django-organizations for 1.7 support for now
git+git://github.com/bennylope/django-organizations.git@dev#egg=django-organizations
django-class-based-auth-views==0.2
arrow==0.4.4
eventlog==0.10.0
future
django-object-actions==0.5.1
django-password-reset==0.7
djangorestframework==3.1.0
drf-extensions==0.2.7
django-rest-swagger==0.2.9
PyCrypto==2.6.1
django-bower==5.0.1
django-sslify>=0.2.4
mailchimp==2.0.9
# django_select2==4.2.2
git+http://github.com/destos/django-select2.git@fix-js-selector#egg=django_select2
pytz==2014.10
django-timezone-field==1.2
django-floppyforms==1.3.0
django-datetime-widget==0.9.3
chardet
django-hijack==1.0.6
simplejson==3.6.5
django-flat-theme==0.9.3
#django-private-media==0.1.3
git+http://github.com/RacingTadpole/django-private-media.git@4ccd2473c3d2c4c61d0747afd1c9c7ae1671737a#egg=private_media
#django-reversion==1.8.5
django-audit-log==0.7.0
django-filter==0.9.2

# for bcrypt passwords
# bcrypt
python-bcrypt

django-ipware==1.1.0
django-maintenancemode==0.11.2

mechanize==0.2.5
django-macaddress==1.3.2

django-bootstrap3==3.3.0
django-bootstrap3-datetimepicker==2.2.3
